break down population report logic into separate class
